+++ dev-tools/dev-server/index.ts
@@ -1,0 +1,8 @@
+import { startDevServer } from './server'
+import { startCoreWatcher } from './core-watcher'
+import { startWebSocketServer } from './web-socket-server'
+
+startDevServer().then(server => {
+  startCoreWatcher()
+  startWebSocketServer(server)
+})