+++ src/components/settings-panel/component-actions/component-actions.ts
@@ -1,5 +1,5 @@
 import { registerAndGetData } from '@/plugins/data'
-import { Executable } from '@/core/common-types'
+import { Executable, VueModule } from '@/core/common-types'
 import { getHook } from '@/plugins/hook'
 import { isUserComponent } from '@/core/settings'
 import { ComponentMetadata } from '../../types'
@@ -10,8 +10,12 @@ export type ComponentAction = (metadata: ComponentMetadata) => {
   displayName: string
   action: Executable
   icon: string
+  visible?: boolean
   title?: string
-  condition?: () => boolean
+  // condition?: () => boolean
+} | {
+  name: string
+  component: Executable<VueModule>
 }
 
 const builtInActions: ComponentAction[] = [
@@ -19,7 +23,7 @@ const builtInActions: ComponentAction[] = [
     name: 'uninstall',
     displayName: '卸载',
     icon: 'mdi-trash-can-outline',
-    condition: () => isUserComponent(metadata),
+    visible: isUserComponent(metadata),
     action: async () => {
       const { before, after } = getHook('userComponents.remove', metadata)
       await before()