+++ src/ui/VButton.vue
@@ -68,6 +68,8 @@ export default Vue.extend({
   &.icon {
     border-radius: 50%;
     padding: 4px;
+    // b 站样式会塞个背景图
+    background-image: none !important;
   }
 
   .content-container {